move avatar handler to business logic
