+++ client/src/business-logic/get-avatar-handler.js
