+++ client/src/business-logic/get-avatar-handler.test.js
